Move Rest classes under appropiate namespace
I also fixed small coding style issues
